+++ klibs/KLParams.py
@@ -54,7 +54,7 @@ blocks_per_experiment = 0
 conditions = []
 default_condition = None
 table_defaults = {} # default column values for db tables when using EntryTemplate
-run_practice_blocks = True # (not implemented in klibs itself)
+run_practice_blocks = True
 color_output = False # whether cso() outputs colorized text or not
 
 # Eye Tracking Settings